rework callbacks and move to own module
move callbacks to own modules in callback folder
