+++ climatefinancebert_ui/components/datatable.py
@@ -5,75 +5,8 @@ from climatefinancebert_ui.components import ids
 
 
 def render(app: Dash):
-    @app.callback(
-        Output(ids.DATATABLE, "children"),
-        [
-            Input("stored-data", "data"),
-            Input(ids.COUNTRIES_LAYER, "clickData"),
-            Input(ids.CATEGORIES_DROPDOWN, "value"),
-            Input(ids.YEAR_SLIDER, "value"),
-        ],
-    )
-    def build_datatable(
-        stored_data=None,
-        click_data=None,
-        selected_categories=None,
-        selected_years=None,
-    ):
-        """
-        Build the datatable based on the input elements and
-        the selected map element.
-        """
-
-        if not click_data or not stored_data:
-            return html.H4("Click a country to render a datatable")
-        else:
-            country_code = click_data["id"]
-            country_name = click_data["properties"]["name"]
-
-            header = [html.H4(f"Data for {country_name}:")]
-
-            df_full = pd.DataFrame(stored_data)
-
-            df_filtered = df_full[
-                (df_full["country_code"] == country_code)
-                & (df_full["meta_category"].isin(selected_categories))
-                & (
-                    df_full["effective_year"].between(
-                        selected_years[0], selected_years[1]
-                    )
-                )
-            ]
-
-            if len(df_filtered.index) == 0:
-                return header + [html.P("No data available")]
-            else:
-                # Render a DataTable with the filtered data
-                return header + [
-                    dash_table.DataTable(
-                        data=df_filtered.to_dict("records"),
-                        columns=[
-                            {
-                                "name": i,
-                                "id": i,
-                                # TODO: style datatable values
-                                # "type": "numeric",
-                                # "format": {"specifier": ".2f"},
-                            }
-                            for i in df_filtered.columns
-                        ],
-                        page_size=15,
-                        sort_action="native",
-                        style_cell={
-                            "overflow": "hidden",
-                            "textOverflow": "ellipsis",
-                            "maxWidth": 0,
-                        },
-                    )
-                ]
-
+    # Return an empty Div that will be filled by the callback
     return html.Div(
-        children=build_datatable(),
         id=ids.DATATABLE,
         className=ids.DATATABLE,
     )
