+++ climatefinancebert_ui/components/ids.py
@@ -10,3 +10,4 @@ INFOBOX = "infobox"
 MAP = "map"
 COUNTRIES_LAYER = "countries"
 DATATABLE = "datatable"
+STORED_DATA = "stored-data"
