+++ climatefinancebert_ui/components/infobox_adaptation.py
@@ -1,29 +1,15 @@
-from dash import Dash, Input, Output, html
+from dash import html
 
 from climatefinancebert_ui.components import ids
 
 
-def render(
-    app: Dash,
-):
-    @app.callback(
-        Output(ids.INFOBOX_ADAPTATION, "children"),
-        Input(ids.YEAR_SLIDER, "value"),
-    )
-    def build_infobox(year=None):
-        header = [html.H5("Adaptation Information")]
-        if not year:
-            return header + [html.P("Select a year")]
-        return header + [
-            html.P(f"Showing data from {year[0]} - {year[1]}"),
-        ]
-
+def render():
+    # Return an empty Div that will be filled by the callback
     return html.Div(
-        children=build_infobox(),
         id=ids.INFOBOX_ADAPTATION,
         className=ids.INFOBOX,
         style={
-            # "position": "absolute",
+            # "position": "absolute"
             "zIndex": "1000",
         },
     )
