+++ climatefinancebert_ui/components/layout.py
@@ -1,24 +1,15 @@
 import dash_bootstrap_components as dbc
-from dash import Dash, Input, Output, dcc, html
+from dash import Dash, dcc, html
 
-from climatefinancebert_ui.components import datatable, ids, map, navbar, sidebar, utils
+from climatefinancebert_ui.components import datatable, ids, map, navbar, sidebar
 
 
 def create_layout(app: Dash) -> html.Div:
-    # Callback to load selected dataset into a shared store
-    @app.callback(
-        Output("stored-data", "data"),
-        [Input(ids.TYPE_DROPDOWN, "value")],
-    )
-    def update_stored_data(selected_type):
-        df_full = utils.fetch_data(selected_type)
-        return df_full.to_dict("records")
-
     return html.Div(
         className="app-container",
         children=[
             navbar.render(app),
-            dcc.Store(id="stored-data"),  # The store to keep the selected dataset
+            dcc.Store(id=ids.STORED_DATA),  # The store to keep the selected dataset
             dbc.Container(
                 fluid=True,  # Set the container to be fluid
                 children=[
