+++ climatefinancebert_ui/components/sidebar.py
@@ -33,14 +33,4 @@ def render(app: Dash):
         ]
     )
 
-    @app.callback(
-        Output("offcanvas", "is_open"),
-        Input("open-offcanvas", "n_clicks"),
-        [State("offcanvas", "is_open")],
-    )
-    def toggle_sidebar(n1, is_open):
-        if n1:
-            return not is_open
-        return is_open
-
     return offcanvas
